加入对byte类型的支持

+++ spring-boot-starter-web-boost-sample/src/main/java/com/fishedee/web_boost/sample/api/User.java
@@ -12,5 +12,5 @@ public class User {
 
     private String name;
 
-    private int age;
+    private byte age;
 }
